requirements.txt
absl-py==2.1.0
accelerate==0.34.2
aiohappyeyeballs==2.4.2
aiohttp==3.10.7
aiosignal==1.3.1
altair==5.4.1
anyio==4.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
crispy-bootstrap4==2024.1
crispy-bootstrap5==2024.2
cycler==0.12.1
dash==2.18.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datasets==3.0.1
debugpy==1.8.3
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
diffusers==0.30.3
dill==0.3.8
dj-database-url==2.2.0
Django==5.1.1
django-admin-interface==0.28.8
django-appconf==1.0.6
django-browser-reload==1.15.0
django-colorfield==0.11.0
django-crispy-forms==2.3
django-forms-dynamic==1.0.0
django-formtools==2.3
django-grappelli==4.0.1
django-htmx==1.19.0
django-import-export==4.1.1
django-select2==8.2.1
django-suit==0.2.28
django-tailwind==3.8.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
et-xmlfile==1.1.0
executing==2.0.1
fastjsonschema==2.20.0
filelock==3.16.1
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.53.1
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.6.1
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.4
gunicorn==23.0.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.7
importlib_metadata==8.4.0
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.4.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
mysqlclient==2.2.4
namex==0.0.8
narwhals==1.8.1
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
notebook==7.2.2
notebook_shim==0.2.4
numpy==2.1.1
openpyxl==3.1.5
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.1
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
patsy==0.5.6
pi==0.1.2
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyarrow==17.0.0
pycparser==2.22
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
python-slugify==8.0.4
pytz==2024.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.5
scapy==2.5.0
scikit-learn==1.5.2
scipy==1.14.0
seaborn==0.13.2
Send2Trash==1.8.3
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
statsmodels==0.14.2
streamlit==1.38.0
sympy==1.13.3
tablib==3.5.0
tenacity==8.5.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.4.0
terminado==0.18.1
text-unidecode==1.3
threadpoolctl==3.5.0
tinycss2==1.3.0
tokenizers==0.20.0
toml==0.10.2
torch==2.4.1
torchvision==0.19.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.45.1
types-python-dateutil==2.9.0.20240906
typing_extensions==4.12.2
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
wcwidth==0.2.13
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
whitenoise==6.7.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xxhash==3.5.0
yarl==1.13.1
zipp==3.20.1
